Introduce `NameMangler.DSL`
